Fix session_id integer overflow in floor_notifications
floor_notifications.session_id was integer but stores a ms timestamp.
Add migration supabase/session_id_bigint.sql to fix.

+++ supabase/dealer_tables.sql
@@ -5,7 +5,7 @@ add column if not exists "tableCount" integer not null default 4;
 -- Уведомления флора (вызовы + выбывания от дилеров)
 create table if not exists public.floor_notifications (
   id uuid primary key default gen_random_uuid(),
-  session_id integer not null,
+  session_id bigint not null,
   type text not null check (type in ('floor_call', 'bustout')),
   table_number integer not null,
   player_id text,
